requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# HTTP client and async support
requests>=2.31.0
aiohttp>=3.9.0
aiofiles>=23.2.0

# Image processing and analysis
pillow>=10.1.0
numpy>=1.24.0

# Environment variable support
python-dotenv>=1.0.0

# Hugging Face dependencies - Enhanced for better models
transformers>=4.42.0
torch>=2.0.0
torchvision>=0.15.0
accelerate>=0.20.0  # For faster model loading

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.1
